feat: outcome invoices, income invoice void/refund, and cross-cutting fixes
Adds branch-expense tracking (Outcome Invoices) and a void/refund flow for
Income Invoices (auto-void when fully refunded). Also folds in numeric-enum
serialization across every service, primary-constructor controllers, an
IAuditable interface so Tenant gets audit stamping, and assorted repository
consolidation (dynamic filters replacing bespoke status params).

+++ Identity.Application/Services/AuthenticationService.cs
@@ -31,7 +31,7 @@ namespace Identity.Application.Services
             };
 
             foreach (var userRole in user.UserRoles)
-                claims.Add(new Claim(ClaimTypes.Role, userRole.RoleId.ToString()));
+                claims.Add(new Claim(ClaimTypes.Role, ((short)userRole.RoleId).ToString()));
 
             foreach (var branch in user.Branches)
             {
